tfm-extract-text-function/requirements.txt
functions-framework==3.*
google-cloud-documentai
google-cloud-storage
